Splitting resolvers and schema
