+++ src/blockscout.ts
@@ -21,7 +21,7 @@ import { Input } from './helpers/Input'
 import { InputDecoder } from './helpers/InputDecoder'
 import { logger } from './logger'
 import { metrics } from './metrics'
-import { TokenTransactionArgs } from './schema'
+import { TokenTransactionArgs } from './resolvers'
 import { Transaction } from './transaction/Transaction'
 import { TransactionAggregator } from './transaction/TransactionAggregator'
 import { TransactionClassifier } from './transaction/TransactionClassifier'
